+++ code/game/objects/items/weapons/storage/fancy.dm
@@ -66,10 +66,12 @@
 	return
 
 /obj/item/storage/fancy/donut_box/decompile_act(obj/item/matter_decompiler/C, mob/user)
-	if(!length(contents))
-		C.stored_comms["wood"] += 1
-		qdel(src)
-		return TRUE
+	if(isdrone(user))
+		if(!length(contents))
+			C.stored_comms["wood"] += 1
+			qdel(src)
+			return TRUE
+		return
 	return ..()
 
 /*
@@ -229,10 +231,12 @@
 	. = ..()
 
 /obj/item/storage/fancy/cigarettes/decompile_act(obj/item/matter_decompiler/C, mob/user)
-	if(!length(contents))
-		C.stored_comms["wood"] += 1
-		qdel(src)
-		return TRUE
+	if(isdrone(user))
+		if(!length(contents))
+			C.stored_comms["wood"] += 1
+			qdel(src)
+			return TRUE
+		return
 	return ..()
 
 /obj/item/storage/fancy/cigarettes/dromedaryco
