+++ code/modules/food_and_drinks/drinks/drinks/bottle.dm
@@ -131,10 +131,12 @@
 	smash(target, thrower, ranged = TRUE)
 
 /obj/item/reagent_containers/food/drinks/bottle/decompile_act(obj/item/matter_decompiler/C, mob/user)
-	if(!reagents.total_volume)
-		C.stored_comms["glass"] += 3
-		qdel(src)
-		return TRUE
+	if(isdrone(user))
+		if(!reagents.total_volume)
+			C.stored_comms["glass"] += 3
+			qdel(src)
+			return TRUE
+		return
 	return ..()
 
 //Keeping this here for now, I'll ask if I should keep it here.
@@ -155,9 +157,11 @@
 	sharp = TRUE
 
 /obj/item/broken_bottle/decompile_act(obj/item/matter_decompiler/C, mob/user)
-	C.stored_comms["glass"] += 3
-	qdel(src)
-	return TRUE
+	if(isdrone(user))
+		C.stored_comms["glass"] += 3
+		qdel(src)
+		return TRUE
+	return ..()
 
 /obj/item/reagent_containers/food/drinks/bottle/gin
 	name = "Griffeater Gin"
